# 47% smaller bundles, 1.8x faster time-to-interactive using dynamic imports