--- a/react/features/base/util/isInsecureRoomName.ts
+++ b/react/features/base/util/isInsecureRoomName.ts
@@ -1,11 +1,11 @@
 import { isEqual } from 'lodash-es';
 import { NIL, parse as parseUUID } from 'uuid';
-import zxcvbn from 'zxcvbn';
 
 // The null UUID.
 const NIL_UUID = parseUUID(NIL);
 
 const _zxcvbnCache = new Map();
+let _zxcvbnPromise: Promise<any> | null = null;
 
 /**
  * Checks if the given string is a valid UUID or not.
@@ -26,16 +26,30 @@ function isValidUUID(str: string) {
 }
 
 /**
+ * Dynamically imports zxcvbn library.
+ *
+ * @returns {Promise} - Promise that resolves to zxcvbn module.
+ */
+function _getZxcvbn() {
+    if (!_zxcvbnPromise) {
+        _zxcvbnPromise = import('zxcvbn').then(module => module.default);
+    }
+
+    return _zxcvbnPromise;
+}
+
+/**
  * Checks a room name and caches the result.
  *
  * @param {string} roomName - The room name.
- * @returns {Object}
+ * @returns {Promise<Object>} - Promise that resolves to zxcvbn result.
  */
-function _checkRoomName(roomName = '') {
+async function _checkRoomName(roomName = '') {
     if (_zxcvbnCache.has(roomName)) {
         return _zxcvbnCache.get(roomName);
     }
 
+    const zxcvbn = await _getZxcvbn();
     const result = zxcvbn(roomName);
 
     _zxcvbnCache.set(roomName, result);
@@ -47,10 +61,15 @@ function _checkRoomName(roomName = '') {
  * Returns true if the room name is considered a weak (insecure) one.
  *
  * @param {string} roomName - The room name.
- * @returns {boolean}
+ * @returns {Promise<boolean>} - Promise that resolves to boolean.
  */
-export default function isInsecureRoomName(roomName = ''): boolean {
-
+export default async function isInsecureRoomName(roomName = ''): Promise<boolean> {
     // room names longer than 200 chars we consider secure
-    return !isValidUUID(roomName) && (roomName.length < 200 && _checkRoomName(roomName).score < 3);
+    if (isValidUUID(roomName) || roomName.length >= 200) {
+        return false;
+    }
+
+    const result = await _checkRoomName(roomName);
+
+    return result.score < 3;
 }