--- a/react/features/conference/components/native/InsecureRoomNameLabel.tsx
+++ b/react/features/conference/components/native/InsecureRoomNameLabel.tsx
@@ -4,26 +4,44 @@ import { connect } from 'react-redux';
 import { translate } from '../../../base/i18n/functions';
 import { IconWarning } from '../../../base/icons/svg';
 import Label from '../../../base/label/components/native/Label';
-import AbstractInsecureRoomNameLabel, { _mapStateToProps } from '../AbstractInsecureRoomNameLabel';
+import useInsecureRoomName from '../../../base/util/useInsecureRoomName';
+import { isUnsafeRoomWarningEnabled } from '../../../prejoin/functions.native';
 
 import styles from './styles';
+
+interface IProps {
+    room: string | undefined;
+    t: Function;
+    unsafeRoomWarningEnabled: boolean;
+}
 
 /**
  * Renders a label indicating that we are in a room with an insecure name.
  */
-class InsecureRoomNameLabel extends AbstractInsecureRoomNameLabel {
-    /**
-     * Renders the platform dependent content.
-     *
-     * @inheritdoc
-     */
-    _render() {
-        return (
-            <Label
-                icon = { IconWarning }
-                style = { styles.insecureRoomNameLabel } />
-        );
+function InsecureRoomNameLabel({ room, unsafeRoomWarningEnabled, t }: IProps) {
+    const isInsecure = useInsecureRoomName(room || '', unsafeRoomWarningEnabled);
+
+    if (!isInsecure) {
+        return null;
     }
+
+    return (
+        <Label
+            icon = { IconWarning }
+            style = { styles.insecureRoomNameLabel } />
+    );
 }
 
-export default translate(connect(_mapStateToProps)(InsecureRoomNameLabel));
+function mapStateToProps(state: any) {
+    const { locked, room } = state['features/base/conference'];
+    const { lobbyEnabled } = state['features/lobby'];
+
+    return {
+        room,
+        unsafeRoomWarningEnabled: Boolean(isUnsafeRoomWarningEnabled(state)
+            && room
+            && !(lobbyEnabled || Boolean(locked)))
+    };
+}
+
+export default translate(connect(mapStateToProps)(InsecureRoomNameLabel));