--- a/react/features/prejoin/components/web/Prejoin.tsx
+++ b/react/features/prejoin/components/web/Prejoin.tsx
@@ -20,7 +20,7 @@ import { getLocalJitsiVideoTrack } from '../../../base/tracks/functions.web';
 import Button from '../../../base/ui/components/web/Button';
 import Input from '../../../base/ui/components/web/Input';
 import { BUTTON_TYPES } from '../../../base/ui/constants.any';
-import isInsecureRoomName from '../../../base/util/isInsecureRoomName';
+import useInsecureRoomName from '../../../base/util/useInsecureRoomName';
 import { openDisplayNamePrompt } from '../../../display-name/actions';
 import { isUnsafeRoomWarningEnabled } from '../../../prejoin/functions';
 import {
@@ -93,6 +93,11 @@ interface IProps {
     readOnlyName: boolean;
 
     /**
+     * The room name to join.
+     */
+    room: string | undefined;
+
+    /**
      * Sets visibility of the 'JoinByPhoneDialog'.
      */
     setJoinByPhoneDialogVisiblity: Function;
@@ -117,15 +122,16 @@ interface IProps {
      */
     showRecordingWarning: boolean;
 
-    /**
-     * If should show unsafe room warning when joining.
-     */
-    showUnsafeRoomWarning: boolean;
 
     /**
      * Whether the user has approved to join a room with unsafe name.
      */
     unsafeRoomConsent?: boolean;
+
+    /**
+     * Whether unsafe room warning is enabled.
+     */
+    unsafeRoomWarningEnabled: boolean;
 
     /**
      * Updates settings.
@@ -221,13 +227,14 @@ const Prejoin = ({
     participantId,
     prejoinConfig,
     readOnlyName,
+    room,
     setJoinByPhoneDialogVisiblity,
     showCameraPreview,
     showDialog,
     showErrorOnJoin,
     showRecordingWarning,
-    showUnsafeRoomWarning,
     unsafeRoomConsent,
+    unsafeRoomWarningEnabled,
     updateSettings: dispatchUpdateSettings,
     videoTrack
 }: IProps) => {
@@ -241,6 +248,9 @@ const Prejoin = ({
     const { classes } = useStyles();
     const { t } = useTranslation();
     const dispatch = useDispatch();
+
+    // Use async hook to check if room name is insecure
+    const showUnsafeRoomWarning = useInsecureRoomName(room || '', unsafeRoomWarningEnabled);
 
     /**
      * Handler for the join button.
@@ -514,12 +524,13 @@ function mapStateToProps(state: IReduxState) {
         participantId,
         prejoinConfig: state['features/base/config'].prejoinConfig,
         readOnlyName: isNameReadOnly(state),
+        room,
         showCameraPreview: !isVideoMutedByUser(state),
         showDialog: isJoinByPhoneDialogVisible(state),
         showErrorOnJoin,
         showRecordingWarning: Boolean(showRecordingWarning),
-        showUnsafeRoomWarning: isInsecureRoomName(room) && isUnsafeRoomWarningEnabled(state),
         unsafeRoomConsent,
+        unsafeRoomWarningEnabled: isUnsafeRoomWarningEnabled(state),
         videoTrack: getLocalJitsiVideoTrack(state)
     };
 }