--- a/react/features/recording/components/Recording/LocalRecordingManager.web.ts
+++ b/react/features/recording/components/Recording/LocalRecordingManager.web.ts
@@ -1,5 +1,3 @@
-// @ts-ignore
-import * as ebml from 'ts-ebml/dist/EBML.min.js';
 import { v4 as uuidV4 } from 'uuid';
 
 import { IStore } from '../../../app/types';
@@ -348,6 +346,9 @@ const LocalRecordingManager: ILocalRecordingManager = {
  * @returns {Promise<Blob>}
  */
 async function fixDuration(data: Blob, duration: number): Promise<Blob> {
+    // Dynamically import ts-ebml library
+    const ebml = await import('ts-ebml/dist/EBML.min.js' as any);
+
     const decoder = new ebml.Decoder();
     const reader = new ebml.Reader();