--- a/react/features/stream-effects/rnnoise/index.ts
+++ b/react/features/stream-effects/rnnoise/index.ts
@@ -1,7 +1,5 @@
 // Script expects to find rnnoise webassembly binary in the same public path root, otherwise it won't load
 // During the build phase this needs to be taken care of manually
-// @ts-expect-error
-import { createRNNWasmModule } from '@jitsi/rnnoise-wasm';
 
 import RnnoiseProcessor from './RnnoiseProcessor';
 
@@ -17,8 +15,12 @@ let rnnoiseModule: Promise<any> | undefined;
  */
 export function createRnnoiseProcessor() {
     if (!rnnoiseModule) {
-        rnnoiseModule = createRNNWasmModule();
+        rnnoiseModule = import('@jitsi/rnnoise-wasm' as any).then(m => {
+            const { createRNNWasmModule } = m;
+
+            return createRNNWasmModule();
+        });
     }
 
-    return rnnoiseModule?.then(mod => new RnnoiseProcessor(mod));
+    return rnnoiseModule.then(mod => new RnnoiseProcessor(mod));
 }