… react/features/conference/components/AbstractInsecureRoomNameLabel.tsx
… react/features/whiteboard/hooks/useExcalidraw.ts