Added dynamic imports for some of the largest libraries:
* rnnoise - Extra noise suppression
* excalidraw - Whiteboard
* zxcvbn/frequency_lists - Secure Room Names
* ts-ebml - Webp Recordings
* amplitude - analytics

Bundle size change:

Metric | Default (Current) | With Dynamic Imports | Improvement
-- | -- | -- | --
Static Size | 11.56 MB | 7.27 MB | -37% (-4.29 MB)
Parsed Size | 4.69 MB | 2.63 MB | -44% (-2.06 MB)
Gzipped Size | 1.49 MB | 788 KB | -47% (-702 KB)

This results in significantly improved application performance across all key metrics, especially noticeable on slower network connections.

**Fast Connection**
Build | DOMContentLoaded | Load | Transferred
-- | -- | -- | --
Default | 869ms | 2.13s | 3.35 MB
Dynamic Imports | 480ms (-45%) | 1.96s | 2.5 MB

**Slow 4G Network (Real-World Conditions)**
Build | DOMContentLoaded | Load | Transferred
-- | -- | -- | --
Default | 11.67s | 20.63s | 3.4 MB
Dynamic Imports | 7.67s (-34%) | 15.85s | 2.5 MB

A live demonstration of the difference in loading times can be seen below:
- [Default (current)](https://sonacove.com/meet/)
- [Updated](https://541d6124-sona-app.catfurr.workers.dev/meet/)

Changes are divided into separate commits to make review easier.
Please let me know if I should update anything!