@saghul 2025-09-15
Exciting! Didn't you need to edit the webpack config too to make them standalone modules?