@catFurr 2025-09-15
Webpack seems to be correctly creating separate chunks for these automatically. I can see that they are also being loaded in the page when needed.
I checked the chunks with the bundle analysis tool.

However, just now realized I might need to edit the Makefile. webpack makes files like `/libs/101.min.js` and `/libs/340.min.js` which need to be copied over.
We don't know in advance what the filename will be, do you have advice on how I can edit the Makefile?