@catFurr 2025-09-30
Will it be possible to get this merged in before the next release?